requirements.txt
attrs==19.1.0
Automat==0.7.0
constantly==15.1.0
cycler==0.10.0
Cython==0.29.13
Django==2.1.1
django-spa==0.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
ninja==1.9.0.post1
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
psycopg2-binary==2.8.3
PyHamcrest==1.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.34.0
whitenoise==3.2.2
yacs==0.1.6
zope.interface==4.6.0
